fix(vz): resolve type mismatch in memory configuration
Cast getDefaultMemoryBytes() result to int to match ProviderConfigInfo.MemoryMB
field type and p.cfg.VZMemoryMB type, fixing compilation errors at lines 528 and 541.

+++ server/internal/sandbox/vz/vz_docker.go
@@ -523,7 +523,7 @@ func (p *DockerProvider) Status() sandbox.ProviderStatus {
 		case DownloadStateReady:
 			status.State = "ready"
 			if kernelPath, baseDiskPath, ok := downloader.GetPaths(); ok {
-				memoryMB := getDefaultMemoryBytes() / (1024 * 1024)
+				memoryMB := int(getDefaultMemoryBytes() / (1024 * 1024))
 				if p.cfg.VZMemoryMB > 0 {
 					memoryMB = p.cfg.VZMemoryMB
 				}
